+++ autorest/codegen/models/lro_operation.py
@@ -10,7 +10,7 @@ from .operation import Operation
 from .parameter import Parameter
 from .schema_response import SchemaResponse
 from .schema_request import SchemaRequest
-from .imports import ImportType
+from .imports import ImportType, TypingSection
 from .object_schema import ObjectSchema
 
 _LOGGER = logging.getLogger(__name__)
@@ -80,9 +80,9 @@ class LROOperation(Operation):
 
     def imports(self, code_model, async_mode: bool) -> FileImport:
         file_import = super().imports(code_model, async_mode)
-        file_import.add_from_import("typing", "Union", ImportType.STDLIB)
+        file_import.add_from_import("typing", "Union", ImportType.STDLIB, TypingSection.CONDITIONAL)
         if async_mode:
-            file_import.add_from_import("typing", "Optional", ImportType.STDLIB)
+            file_import.add_from_import("typing", "Optional", ImportType.STDLIB, TypingSection.CONDITIONAL)
             file_import.add_from_import("azure.core.polling", "async_poller", ImportType.AZURECORE)
             file_import.add_from_import("azure.core.polling", "AsyncNoPolling", ImportType.AZURECORE)
             file_import.add_from_import("azure.core.polling", "AsyncPollingMethod", ImportType.AZURECORE)